chore: lazy load RichTextMenu
