+++ packages/core/components/RichTextMenu/controls/HeadingSelect/index.tsx
@@ -1,0 +1,14 @@
+import { lazy, Suspense } from "react";
+import { HeadingSelectFallback } from "./fallback";
+
+const HeadingSelectLoaded = lazy(() =>
+  import("./loaded").then((m) => ({
+    default: m.HeadingSelectLoaded,
+  }))
+);
+
+export const HeadingSelect = () => (
+  <Suspense fallback={<HeadingSelectFallback />}>
+    <HeadingSelectLoaded />
+  </Suspense>
+);
