+++ packages/core/components/RichTextMenu/index.tsx
@@ -1,51 +1,25 @@
-import { useEditorState } from "@tiptap/react";
-import { useMemo } from "react";
-import {
-  EditorState,
-  RichTextEditor,
-  RichTextSelector,
-} from "../RichTextEditor/types";
-import { defaultEditorState } from "../RichTextEditor/selector";
+import { lazy, Suspense } from "react";
+import { EditorState, RichTextEditor } from "../RichTextEditor/types";
 import { RichtextField } from "../../types";
 import { LoadedRichTextMenuInner } from "./inner";
 
-export const LoadedRichTextMenu = ({
-  editor,
-  field,
-  readOnly,
-  inline,
-}: {
+const LoadedRichTextMenuFull = lazy(() =>
+  import("./full").then((m) => ({
+    default: m.LoadedRichTextMenuFull,
+  }))
+);
+
+export type LoadedRichTextMenuProps = {
   field: RichtextField;
   editor: RichTextEditor | null;
+  editorState?: EditorState | null;
   readOnly: boolean;
   inline?: boolean;
-}) => {
-  const { tiptap = {} } = field;
-  const { selector } = tiptap;
-
-  const resolvedSelector = useMemo(() => {
-    return (ctx: Parameters<RichTextSelector>[0]) => ({
-      ...defaultEditorState(ctx, readOnly),
-      ...(selector ? selector(ctx, readOnly) : {}),
-    });
-  }, [selector, readOnly]);
-
-  const editorState = useEditorState<EditorState>({
-    editor,
-    selector: resolvedSelector,
-  });
-
-  if (!editor || !editorState) {
-    return null;
-  }
-
+};
+export const LoadedRichTextMenu = (props: LoadedRichTextMenuProps) => {
   return (
-    <LoadedRichTextMenuInner
-      editor={editor}
-      editorState={editorState}
-      field={field}
-      readOnly={readOnly}
-      inline={inline}
-    />
+    <Suspense fallback={<LoadedRichTextMenuInner {...props} />}>
+      <LoadedRichTextMenuFull {...props} />
+    </Suspense>
   );
 };
